Update tournament view

+++ src/comp3004/ivanhoe/util/ClientParser.java
@@ -57,8 +57,8 @@ public class ClientParser {
 		return (String)((JSONObject)player).get("display_total");
 	}
 	
-	public Long getPlayerId(Object player) {
-		return (Long)((JSONObject)player).get("id");
+	public Integer getPlayerId(Object player) {
+		return (Integer)((JSONObject)player).get("id");
 	}
 	
 	@SuppressWarnings("unchecked")
